backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
PyYAML>=6.0
python-multipart>=0.0.9
